refactor: page 컴포넌트 비즈니스 로직 hook으로 분리
