+++ src/pages/CartPage/index.tsx
@@ -1,22 +1,11 @@
 import PaymentsAmount from 'components/Cart/PaymentsAmount';
-import { useEffect } from 'react';
-import { useAppDispatch } from 'hooks/useAppDispatch';
-import { useAppSelector } from 'hooks/useAppSelector';
-import { getCartList } from 'redux/action-creators/cartListThunk';
-import { CartListAction } from 'redux/actions/cartList';
 import CartList from 'components/Cart/CartList';
 import RequestFail from 'components/@common/RequestFail';
 import { Styled } from './styles';
+import { useCart } from './useCart';
 
 const CartPage = () => {
-  const { data: cartList, error: errorGetCartList } = useAppSelector(
-    state => state.cartListReducer
-  );
-  const dispatch = useAppDispatch<CartListAction>();
-
-  useEffect(() => {
-    dispatch(getCartList());
-  }, []);
+  const { cartList, errorGetCartList } = useCart();
 
   if (errorGetCartList) return <RequestFail />;
