+++ src/pages/EditPasswordPage/index.tsx
@@ -1,48 +1,18 @@
-import { FormEvent } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { editPassword } from 'redux/action-creators/userThunk';
-import { UserAction } from 'redux/actions/user';
-import { useAppDispatch } from 'hooks/useAppDispatch';
-import { useAppSelector } from 'hooks/useAppSelector';
-import usePasswordInput from 'hooks/usePasswordInput';
 import SignInput from 'components/@common/SignInput';
-import { PATH } from 'Router';
-import { useDispatch } from 'react-redux';
-import { updateSnackBar } from 'redux/actions/snackBar';
-import { MESSAGE } from 'constant/message';
 import { Styled } from './styles';
+import { useEditPassword } from './useEditPassword';
 
 const EditPasswordPage = () => {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const thunkDispatch = useAppDispatch<UserAction>();
-  const { data } = useAppSelector(state => state.userReducer);
   const {
+    data,
+    passwordValid,
     prevPasswordRef,
     currentPasswordRef,
-    passwordValid,
+    handleSubmit,
     handlePrevPasswordInput,
     handleCurrentPasswordInput,
     handleCurrentPasswordConfirmInput,
-  } = usePasswordInput();
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const editPasswordInfo = {
-      password: prevPasswordRef.current.value,
-      newPassword: currentPasswordRef.current.value,
-    };
-
-    if (Object.values(passwordValid).every(valid => valid)) {
-      try {
-        await thunkDispatch(editPassword(editPasswordInfo));
-        navigate(PATH.default);
-      } catch (error) {
-        dispatch(updateSnackBar(MESSAGE.FAILED_EDIT_PASSWORD));
-      }
-    }
-  };
+  } = useEditPassword();
 
   return (
     <Styled.EditPasswordPage>
