+++ src/pages/ItemListPage/index.tsx
@@ -1,25 +1,11 @@
-import { useEffect } from 'react';
-import { getItemList } from 'redux/action-creators/itemListThunk';
-import { ItemListAction } from 'redux/actions/itemList';
-import { useAppSelector } from 'hooks/useAppSelector';
-import { useAppDispatch } from 'hooks/useAppDispatch';
-import { getCartList } from 'redux/action-creators/cartListThunk';
 import ItemList from 'components/ItemList';
 import Paginator from 'components/@common/Paginator';
 import Loading from 'components/@common/Loading';
 import RequestFail from 'components/@common/RequestFail';
-import { getLocalStorageToken } from 'utils/localStorage';
+import { useItemList } from './useItemList';
 
 const ItemListPage = () => {
-  const dispatch = useAppDispatch<ItemListAction>();
-  const { data: itemList, loading, error } = useAppSelector(state => state.itemListReducer);
-
-  useEffect(() => {
-    dispatch(getItemList());
-    if (getLocalStorageToken()) {
-      dispatch(getCartList());
-    }
-  }, []);
+  const { loading, error, itemList } = useItemList();
 
   if (loading) return <Loading />;
   if (error) return <RequestFail />;
