+++ src/pages/SignUpPage/index.tsx
@@ -1,50 +1,19 @@
-import { FormEvent } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { signUp } from 'redux/action-creators/userThunk';
-import { UserAction } from 'redux/actions/user';
-import { updateSnackBar } from 'redux/actions/snackBar';
-import { useAppDispatch } from 'hooks/useAppDispatch';
-import useSignInput from 'hooks/useSignInput';
-import usePasswordInput from 'hooks/usePasswordInput';
 import SignInput from 'components/@common/SignInput';
-import { PATH } from 'Router';
-import { useDispatch } from 'react-redux';
-import { MESSAGE } from 'constant/message';
+
 import { Styled } from './styles';
+import { useSignUp } from './useSignUp';
 
 const SignUpPage = () => {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const thunkDispatch = useAppDispatch<UserAction>();
-  const { inputState, validState, handleEmailInput, handleNameInput } = useSignInput();
   const {
+    validState,
     currentPasswordRef,
     passwordValid,
+    handleEmailInput,
+    handleNameInput,
     handleCurrentPasswordInput,
     handleCurrentPasswordConfirmInput,
-  } = usePasswordInput();
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const inputInfo = {
-      email: inputState.email,
-      name: inputState.name,
-      password: currentPasswordRef.current.value,
-    };
-
-    if (
-      Object.values(validState).every(valid => valid) &&
-      [passwordValid.current, passwordValid.confirm].every(valid => valid)
-    ) {
-      try {
-        await thunkDispatch(signUp(inputInfo));
-        navigate(PATH.signIn);
-      } catch (error) {
-        dispatch(updateSnackBar(MESSAGE.FAILED_SIGN_UP));
-      }
-    }
-  };
+    handleSubmit,
+  } = useSignUp();
 
   return (
     <Styled.SignUpPage>
